3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -25,7 +25,7 @@ void print_all(const char * const format, ...)
 				check_stat = 0; /* CHECKS CONDITION */
 				break;
 			case 'f':
-				printf("%f", va_arg(spc, double));
+				printf("%f", va_arg(spc, float));
 				check_stat = 0;
 				break;
 			case 'c':
